requirements.txt
asgiref==3.2.10
autopep8==1.5.4
defusedxml==0.6.0
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-xml==2.0.0
mysqlclient==2.0.1
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
